+++ src/backend/modules/value/ValueDTO.ts
@@ -1,0 +1,8 @@
+export class ValueDTO {
+
+  id: number;
+  value: string;
+}
+
+export type CreateValueDTO = Pick<ValueDTO, 'value'>
+export type UpdateValueDTO = Pick<ValueDTO, 'value'>